Add NTS Radio
Add support for play / pause / toggle on NTS Radio's website
(https://nts.live).

Note: When listening to audio from NTS.live, it's from one of two
sources:

    * An <audio> element
    * An <iframe> with a Mixcloud widget inside.

The first source is easy enough to work with, but the mixcloud one is
not. While Mixcloud does provide an API to interact with it, it's
entirely asynchronous. This is fine for playing / pausing, but if we
want to retrieve any information from it, further hacks / support will
need to be implemented.